+++ island-simcity/src/main/java/com/island/simcity/entities/components/BuildingComponent.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @Setter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class BuildingComponent implements Component {
@@ -25,26 +26,13 @@ public class BuildingComponent implements Component {
         POWER_LINE
     }
 
+    public enum Density {
+        LOW,
+        MEDIUM,
+        HIGH
+    }
+
     private Type type;
-
-    public Type getType() {
-        return type;
-    }
-
-    public static BuildingComponentBuilder builder() {
-        return new BuildingComponentBuilder();
-    }
-
-    public static class BuildingComponentBuilder {
-        private Type type;
-
-        public BuildingComponentBuilder type(Type type) {
-            this.type = type;
-            return this;
-        }
-
-        public BuildingComponent build() {
-            return new BuildingComponent(type);
-        }
-    }
+    @Builder.Default
+    private Density density = Density.LOW;
 }